+++ src/Launcher.ts
@@ -23,7 +23,8 @@ import * as fs from 'fs';
 import { BrowserFetcher } from './BrowserFetcher';
 import { Connection } from './Connection';
 import { Browser } from './Browser';
-import { helper, assert, debugError } from './helper';
+import { assert } from './assert';
+import { helper, debugError } from './helper';
 import { ConnectionTransport } from './ConnectionTransport';
 import { WebSocketTransport } from './WebSocketTransport';
 import { BrowserRunner } from './launcher/BrowserRunner';
