+++ src/Page.ts
@@ -24,7 +24,8 @@ import { EmulationManager } from './EmulationManager';
 import { Frame, FrameManager } from './FrameManager';
 import { Keyboard, Mouse, Touchscreen, MouseButtonInput } from './Input';
 import { Tracing } from './Tracing';
-import { helper, debugError, assert } from './helper';
+import { assert } from './assert';
+import { helper, debugError } from './helper';
 import { Coverage } from './Coverage';
 import { WebWorker } from './WebWorker';
 import { Browser, BrowserContext } from './Browser';
