+++ src/Tracing.ts
@@ -13,7 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { helper, assert } from './helper';
+import { assert } from './assert';
+import { helper } from './helper';
 import { CDPSession } from './Connection';
 
 interface TracingOptions {
